--- a/regression-test/data/nereids_tpcds_shape_sf100_p0/noStatsRfPrune/query44.out
+++ b/regression-test/data/nereids_tpcds_shape_sf100_p0/noStatsRfPrune/query44.out
@@ -12,33 +12,32 @@ PhysicalResultSink
 ------------------PhysicalDistribute
 --------------------PhysicalProject
 ----------------------hashJoin[INNER_JOIN] hashCondition=((asceding.rnk = descending.rnk)) otherCondition=()
-------------------------PhysicalDistribute
---------------------------PhysicalProject
-----------------------------filter((rnk < 11))
-------------------------------PhysicalWindow
---------------------------------PhysicalQuickSort[MERGE_SORT]
-----------------------------------PhysicalDistribute
-------------------------------------PhysicalQuickSort[LOCAL_SORT]
---------------------------------------PhysicalPartitionTopN
-----------------------------------------PhysicalProject
-------------------------------------------NestedLoopJoin[INNER_JOIN](cast(rank_col as DOUBLE) > cast((0.9 * rank_col) as DOUBLE))
---------------------------------------------PhysicalProject
-----------------------------------------------hashAgg[GLOBAL]
-------------------------------------------------PhysicalDistribute
---------------------------------------------------hashAgg[LOCAL]
-----------------------------------------------------PhysicalProject
-------------------------------------------------------filter((ss1.ss_store_sk = 146))
---------------------------------------------------------PhysicalOlapScan[store_sales]
---------------------------------------------PhysicalDistribute
-----------------------------------------------PhysicalAssertNumRows
-------------------------------------------------PhysicalDistribute
+------------------------PhysicalProject
+--------------------------filter((rnk < 11))
+----------------------------PhysicalWindow
+------------------------------PhysicalQuickSort[MERGE_SORT]
+--------------------------------PhysicalDistribute
+----------------------------------PhysicalQuickSort[LOCAL_SORT]
+------------------------------------PhysicalPartitionTopN
+--------------------------------------PhysicalProject
+----------------------------------------NestedLoopJoin[INNER_JOIN](cast(rank_col as DOUBLE) > cast((0.9 * rank_col) as DOUBLE))
+------------------------------------------PhysicalProject
+--------------------------------------------hashAgg[GLOBAL]
+----------------------------------------------PhysicalDistribute
+------------------------------------------------hashAgg[LOCAL]
 --------------------------------------------------PhysicalProject
-----------------------------------------------------hashAgg[GLOBAL]
-------------------------------------------------------PhysicalDistribute
---------------------------------------------------------hashAgg[LOCAL]
-----------------------------------------------------------PhysicalProject
-------------------------------------------------------------filter((store_sales.ss_store_sk = 146) and ss_addr_sk IS NULL)
---------------------------------------------------------------PhysicalOlapScan[store_sales]
+----------------------------------------------------filter((ss1.ss_store_sk = 146))
+------------------------------------------------------PhysicalOlapScan[store_sales]
+------------------------------------------PhysicalDistribute
+--------------------------------------------PhysicalAssertNumRows
+----------------------------------------------PhysicalDistribute
+------------------------------------------------PhysicalProject
+--------------------------------------------------hashAgg[GLOBAL]
+----------------------------------------------------PhysicalDistribute
+------------------------------------------------------hashAgg[LOCAL]
+--------------------------------------------------------PhysicalProject
+----------------------------------------------------------filter((store_sales.ss_store_sk = 146) and ss_addr_sk IS NULL)
+------------------------------------------------------------PhysicalOlapScan[store_sales]
 ------------------------PhysicalDistribute
 --------------------------PhysicalProject
 ----------------------------filter((rnk < 11))