… regression-test/data/nereids_tpcds_shape_sf100_p0/noStatsRfPrune/query54.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/noStatsRfPrune/query71.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/no_stats_shape/query44.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/no_stats_shape/query5.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/no_stats_shape/query54.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/no_stats_shape/query71.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/no_stats_shape/query74.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/rf_prune/query44.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/rf_prune/query74.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/rf_prune/query78.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query44.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query74.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query78.out