## Proposed changes
1. data skew reduces shuffle join parallelism, where broadcast join is a better choice.
2. fix a bug in union column stats. when any child of union is unknown, union stats should be unknown. tpcds 74 is affected by this fix. but there is no impact on performance 
ds78 变差还需调查
before: 
query44 4358    1073    876     876
query74 6151    4995    4888    4888
query78 7655    5637    5704    5637
Total cold run time: 18164 ms
Total hot run time: 11401 ms

after: 
query44 4479    1066    781     781
query74 5851    4929    4830    4830
query78 12365   10523   10689   10523
Total cold run time: 22695 ms
Total hot run time: 16134 ms
Issue Number: close #xxx

<!--Describe your changes.-->

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...