@englefly 2023-11-28
run buildall