@englefly 2023-11-29
run buildall